+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.tests.js
@@ -65,6 +65,7 @@ import { DataSource } from 'data/data_source/data_source';
 import ArrayStore from 'data/array_store';
 import messageLocalization from 'localization/message';
 import { setTemplateEngine } from 'core/templates/template_engine_registry';
+import gridCoreUtils from 'ui/grid_core/ui.grid_core.utils';
 import fx from 'animation/fx';
 import config from 'core/config';
 import keyboardMock from '../../helpers/keyboardMock.js';
@@ -15920,6 +15921,48 @@ QUnit.module('API methods', baseModuleConfig, () => {
             'rowExpanded rowCount: 2, masterRow: true, detailRow: true',
             'done rowCount: 2, masterRow: true, detailRow: true'
         ]);
+    });
+
+    QUnit.testInActiveWindow('Filter row editor should have focus after _synchronizeColumns (T638737)', function(assert) {
+        $('#qunit-fixture').css('position', 'static');
+        // arrange, act
+        const dataGrid = createDataGrid({
+            filterRow: { visible: true },
+            editing: { allowAdding: true },
+            columns: [
+                { dataField: 'field1' },
+                { dataField: 'field2' }
+            ],
+            dataSource: [{ field1: 1, field2: 2 }, { field1: 3, field2: 4 }]
+        });
+
+        this.clock.tick();
+
+        const $input = $(dataGrid.$element()).find('.dx-editor-cell').first().find('.dx-texteditor-input');
+        $input.focus().val('1').trigger('change');
+
+        const selectionRangeArgs = [];
+
+        const oldSetSelectionRange = gridCoreUtils.setSelectionRange;
+        gridCoreUtils.setSelectionRange = function(element, range) {
+            oldSetSelectionRange.apply(this, arguments);
+            selectionRangeArgs.push([element, range]);
+        };
+
+        this.clock.tick();
+
+        gridCoreUtils.setSelectionRange = oldSetSelectionRange;
+
+        // assert
+        const $focusedInput = dataGrid.$element().find('.dx-editor-cell .dx-texteditor-input:focus');
+        assert.equal(dataGrid.getVisibleRows().length, 1, 'filter was applied');
+        assert.ok($focusedInput.length, 'filter cell has focus after filter applyed');
+        // T662207
+        if(devices.real().deviceType === 'desktop') {
+            assert.deepEqual(selectionRangeArgs, [[$focusedInput.get(0), { selectionStart: 1, selectionEnd: 1 }]], 'setSelectionRange args');
+        }
+
+        $('#qunit-fixture').css('position', '');
     });
 });
